+++ wasmbinding/exec_test.go
@@ -98,9 +98,11 @@ func (s *TestSuiteExecutor) SetupSuite() {
 		ChainID: "nibiru-wasmnet-1",
 		Time:    time.Now().UTC(),
 	})
+	nibiru.PerpKeeperV2.Collateral.Set(ctx, perpv2types.TestingCollateralDenomNUSD)
+
 	coins := sdk.NewCoins(
 		sdk.NewCoin(denoms.NIBI, sdk.NewInt(10_000_000)),
-		sdk.NewCoin(denoms.NUSD, sdk.NewInt(420_000*69)),
+		sdk.NewCoin(perpv2types.TestingCollateralDenomNUSD, sdk.NewInt(420_000*69)),
 	)
 
 	s.NoError(testapp.FundAccount(nibiru.BankKeeper, ctx, sender, coins))
@@ -110,6 +112,7 @@ func (s *TestSuiteExecutor) SetupSuite() {
 	s.ctx = ctx
 	s.keeper = TestOnlySudoKeeper{Keeper: s.nibiru.SudoKeeper}
 	s.wasmKeeper = wasmkeeper.NewDefaultPermissionKeeper(nibiru.WasmKeeper)
+	s.nibiru.PerpKeeperV2.Collateral.Set(s.ctx, perpv2types.TestingCollateralDenomNUSD)
 
 	s.contractPerp = ContractMap[wasmbin.WasmKeyPerpBinding]
 	s.contractController = ContractMap[wasmbin.WasmKeyController]
@@ -126,10 +129,10 @@ func (s *TestSuiteExecutor) OnSetupEnd() {
 
 func (s *TestSuiteExecutor) TestOpenAddRemoveClose() {
 	pair := asset.MustNewPair(s.happyFields.Pair)
-	margin := sdk.NewCoin(denoms.NUSD, sdk.NewInt(69))
+	margin := sdk.NewCoin(perpv2types.TestingCollateralDenomNUSD, sdk.NewInt(69))
 
 	coins := sdk.NewCoins(
-		margin.Add(sdk.NewCoin(denoms.NUSD, sdk.NewInt(1_000))),
+		margin.Add(sdk.NewCoin(perpv2types.TestingCollateralDenomNUSD, sdk.NewInt(1_000))),
 	)
 	s.NoError(testapp.FundAccount(s.nibiru.BankKeeper, s.ctx, s.contractPerp, coins))
 
@@ -466,7 +469,7 @@ func (s *TestSuiteExecutor) TestInsuranceFundWithdraw() {
 		s.nibiru.BankKeeper,
 		s.ctx,
 		perpv2types.PerpEFModuleAccount,
-		sdk.NewCoins(sdk.NewCoin(denoms.NUSD, sdk.NewInt(420))),
+		sdk.NewCoins(sdk.NewCoin(perpv2types.TestingCollateralDenomNUSD, sdk.NewInt(420))),
 	)
 	s.NoError(err)
 	s.keeper.SetSudoContracts(